refactor(release): split manifest.ts CLI layer into manifest-cli.ts (#691)
Lift HELP_TEXT, the flag grammar, run, the check/emit helpers, and the
report renderers out of manifest.ts into a sibling manifest-cli.ts,
leaving manifest.ts as the classifier, exclude parsing, build, lints,
and drift engine. Behavior identical; computeMissing/computeMissingEntries
stay the single shared extraction point and computeDrift stays 3-arg.
Tests split to mirror the modules (73 passed + 1 skipped in each half,
matching the original manifest.test.ts total of 74).

manifest.ts is now 520 lines (down from 1331) but still exceeds the
~400-line guideline; manifest-cli.ts is 850 lines because the flag
grammar plus emit/run layer is inherently large. Both pass lint clean,
including sonarjs/cognitive-complexity, with no per-function complexity
trips. Fully honoring the ~400-line guideline would need a second split
within the CLI layer, out of scope for this pure-refactor issue.

Closes #691
